+++ product_category_attributes/models/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from . import product_category, product_template
+from . import product_category, product_category_attribute_line, product_template
